requirements.txt
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
torch>=2.0.0
transformers>=4.35.0

# DL perception (CLIP) — install separately if you have GPU/enough RAM:
# torch>=2.0.0
# transformers>=4.35.0
